requirements.txt
# requirements.txt - Based on your working Python 3.11.2 environment

# Financial data
yfinance==0.2.33
sec-api==1.0.17

# Google Cloud basics
google-auth==2.39.0
google-cloud-secret-manager==2.24.0
google-cloud-logging==3.11.4

# Web Framework
fastapi==0.115.12
uvicorn==0.34.3

# HTTP requests and Web Scraping
requests==2.32.4
beautifulsoup4==4.13.4

# Data processing
pandas>=2.0.0
numpy>=2.0.0

# Database (your working versions)
sqlalchemy==2.0.41
pg8000==1.31.2
cloud-sql-python-connector==1.18.2

# Core Google ADK and AI Platform (your working versions)
google-cloud-aiplatform==1.97.0
google-adk==1.3.0
vertexai  # Will use latest compatible

# Environment and Utilities
python-dotenv==1.1.0
python-dateutil==2.9.0.post0

# FastAPI Dependencies
pydantic==2.11.7
python-multipart==0.0.20

# Security
cryptography==45.0.3

# Additional dependencies that might be needed
google-api-core==2.24.2
google-cloud-core==2.4.3
googleapis-common-protos==1.70.0
grpcio==1.71.0
protobuf==5.29.4
